remaned K2Raster files changed RavenJoin so it supports long k2-raster trees

+++ raven/src/main/java/dk/itu/raven/join/RavenJoin.java
@@ -194,17 +194,17 @@ public class RavenJoin {
 	 *         k2 raster tree</li>
 	 *         </ul>
 	 */
-	private Tuple5<QuadOverlapType, Integer, Square, Integer, Integer> checkQuadrant(int k2Index, Square rasterBounding,
-			Rectangle bounding, int lo, int hi, int min, int max) {
-		int vMinMBR = min;
-		int vMaxMBR = max;
+	private Tuple5<QuadOverlapType, Integer, Square, Long, Long> checkQuadrant(int k2Index, Square rasterBounding,
+			Rectangle bounding, long lo, long hi, long min, long max) {
+		long vMinMBR = min;
+		long vMaxMBR = max;
 		Logger.log(vMinMBR + ", " + vMaxMBR);
 		int returnedK2Index = k2Index;
 		Square returnedrasterBounding = rasterBounding;
-		Stack<Tuple4<Integer, Square, Integer, Integer>> k2Nodes = new Stack<>();
+		Stack<Tuple4<Integer, Square, Long, Long>> k2Nodes = new Stack<>();
 		k2Nodes.push(new Tuple4<>(k2Index, rasterBounding, min, max));
 		while (!k2Nodes.empty()) {
-			Tuple4<Integer, Square, Integer, Integer> node = k2Nodes.pop();
+			Tuple4<Integer, Square, Long, Long> node = k2Nodes.pop();
 			int[] children = k2Raster.getChildren(node.a);
 			int childSize = node.b.getSize() / k2Raster.k;
 			for (int i = 0; i < children.length; i++) {
@@ -248,15 +248,15 @@ public class RavenJoin {
 	 * @return one of {@code TotalOverlap, PartialOverlap, NoOverlap}
 	 */
 	private MBROverlapType checkMBR(int k2Index, Square rasterBounding, Rectangle bounding,
-			int lo, int hi, int min, int max) {
-		int vMinMBR = Integer.MAX_VALUE;
-		int vMaxMBR = Integer.MIN_VALUE;
+			long lo, long hi, long min, long max) {
+		long vMinMBR = Long.MAX_VALUE;
+		long vMaxMBR = Long.MIN_VALUE;
 
-		Stack<Tuple4<Integer, Square, Integer, Integer>> k2Nodes = new Stack<>();
+		Stack<Tuple4<Integer, Square, Long, Long>> k2Nodes = new Stack<>();
 		k2Nodes.push(new Tuple4<>(k2Index, rasterBounding, min, max));
 
 		while (!k2Nodes.empty()) {
-			Tuple4<Integer, Square, Integer, Integer> node = k2Nodes.pop();
+			Tuple4<Integer, Square, Long, Long> node = k2Nodes.pop();
 			int[] children = k2Raster.getChildren(node.a);
 			int childSize = node.b.getSize() / k2Raster.k;
 
@@ -270,8 +270,8 @@ public class RavenJoin {
 				Square childRasterBounding = node.b.getChildSquare(childSize, i, k2Raster.k);
 
 				if (childRasterBounding.intersects(bounding)) {
-					int vminVal = k2Raster.computeVMin(node.d, node.c, child);
-					int vmaxVal = k2Raster.computeVMax(node.d, child);
+					long vminVal = k2Raster.computeVMin(node.d, node.c, child);
+					long vmaxVal = k2Raster.computeVMax(node.d, child);
 					if (childRasterBounding.isContained(bounding)) {
 						vMinMBR = Math.min(vminVal, vMinMBR);
 						vMaxMBR = Math.max(vmaxVal, vMaxMBR);
@@ -281,7 +281,7 @@ public class RavenJoin {
 				}
 			}
 		}
-		if (vMinMBR == Integer.MAX_VALUE || vMaxMBR == Integer.MIN_VALUE) {
+		if (vMinMBR == Long.MAX_VALUE || vMaxMBR == Long.MIN_VALUE) {
 			throw new RuntimeException("rasterBounding was never contained in bounding");
 		}
 		if (vMinMBR >= lo && vMaxMBR <= hi) {
@@ -300,7 +300,7 @@ public class RavenJoin {
 	 *         it contains
 	 */
 	public List<Pair<Geometry, Collection<PixelRange>>> join() {
-		return join(Integer.MIN_VALUE, Integer.MAX_VALUE);
+		return join(Long.MIN_VALUE, Long.MAX_VALUE);
 	}
 
 	// based on:
@@ -313,20 +313,20 @@ public class RavenJoin {
 	 * @return a list of Geometries paired with a collection of the pixelranges,
 	 *         whose values fall within the given range, that it contains
 	 */
-	public List<Pair<Geometry, Collection<PixelRange>>> join(int lo, int hi) {
+	public List<Pair<Geometry, Collection<PixelRange>>> join(long lo, long hi) {
 		List<Pair<Geometry, Collection<PixelRange>>> def = new ArrayList<>(), prob = new ArrayList<>();
-		Stack<Tuple5<Node<String, Geometry>, Integer, Square, Integer, Integer>> S = new Stack<>();
+		Stack<Tuple5<Node<String, Geometry>, Integer, Square, Long, Long>> S = new Stack<>();
 
-		int[] minMax = k2Raster.getValueRange();
+		long[] minMax = k2Raster.getValueRange();
 
 		for (Node<String, Geometry> node : TreeExtensions.getChildren(tree.root().get())) {
 			S.push(new Tuple5<>(node, 0, new Square(0, 0, k2Raster.getSize()), minMax[0], minMax[1]));
 		}
 
 		while (!S.empty()) {
-			Tuple5<Node<String, Geometry>, Integer, Square, Integer, Integer> p = S.pop();
+			Tuple5<Node<String, Geometry>, Integer, Square, Long, Long> p = S.pop();
 			if (!new Square(0, 0, k2Raster.getSize()).intersects(p.a.geometry().mbr())) continue;
-			Tuple5<QuadOverlapType, Integer, Square, Integer, Integer> checked = checkQuadrant(p.b, p.c, p.a.geometry().mbr(),
+			Tuple5<QuadOverlapType, Integer, Square, Long, Long> checked = checkQuadrant(p.b, p.c, p.a.geometry().mbr(),
 					lo, hi, p.d,
 					p.e);
 			switch (checked.a) {
@@ -340,7 +340,7 @@ public class RavenJoin {
 				case PossibleOverlap:
 					if (!TreeExtensions.isLeaf(p.a)) {
 						for (Node<String, Geometry> c : ((NonLeaf<String, Geometry>) p.a).children()) {
-							S.push(new Tuple5<Node<String, Geometry>, Integer, Square, Integer, Integer>(c, checked.b, checked.c,
+							S.push(new Tuple5<Node<String, Geometry>, Integer, Square, Long, Long>(c, checked.b, checked.c,
 									checked.d, checked.e));
 						}
 					} else {
@@ -379,12 +379,12 @@ public class RavenJoin {
 	 * @param hi   the maximum pixel-value that should be included in the join
 	 */
 	protected void combineLists(List<Pair<Geometry, Collection<PixelRange>>> def,
-			List<Pair<Geometry, Collection<PixelRange>>> prob, int lo, int hi) {
+			List<Pair<Geometry, Collection<PixelRange>>> prob, long lo, long hi) {
 		Logger.log("def: " + def.size() + ", prob: " + prob.size());
 		for (Pair<Geometry, Collection<PixelRange>> pair : prob) {
 			Pair<Geometry, Collection<PixelRange>> result = new Pair<>(pair.first, new ArrayList<>());
 			for (PixelRange range : pair.second) {
-				int[] values = k2Raster.getWindow(range.row, range.row, range.x1, range.x2);
+				long[] values = k2Raster.getWindow(range.row, range.row, range.x1, range.x2);
 				for (int i = 0; i < values.length; i++) {
 					int start = i;
 					while (i < values.length && values[i] >= lo && values[i] <= hi) {
