+++ raven/src/test/java/dk/itu/raven/ksquared/KSquaredTest.java
@@ -1,7 +1,6 @@
 package dk.itu.raven.ksquared;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
@@ -11,14 +10,13 @@ import java.io.ObjectOutputStream;
 import java.util.Random;
 import java.util.Stack;
 
-import org.junit.jupiter.api.Test;
-import org.geotools.api.metadata.citation.Address;
 import org.junit.jupiter.api.RepeatedTest;
+import org.junit.jupiter.api.Test;
 
-import dk.itu.raven.util.matrix.RandomMatrix;
 import dk.itu.raven.join.Square;
 import dk.itu.raven.util.matrix.ArrayMatrix;
 import dk.itu.raven.util.matrix.Matrix;
+import dk.itu.raven.util.matrix.RandomMatrix;
 
 public class KSquaredTest {
 	private final int[][] M = {
@@ -39,7 +37,7 @@ public class KSquaredTest {
 		int row = r.nextInt(100);
 		int col1 = r.nextInt(25);
 		int col2 = 75 + r.nextInt(25);
-		int[] res = k2Raster.getWindow(row, row, col1, col2);
+		long[] res = k2Raster.getWindow(row, row, col1, col2);
 		for (int i = 0; i < res.length; i++) {
 			assertEquals(res[i], matrix.get(row, col1 + i));
 		}
@@ -57,7 +55,7 @@ public class KSquaredTest {
 		Matrix matrix = new RandomMatrix(2000, 500, 1000000);
 		K2Raster k2Raster = new K2RasterBuilder().build(matrix,2);
 		for (int i = 0; i < matrix.getHeight(); i++) {
-			int[] row = k2Raster.getWindow(i, i, 0, matrix.getWidth() - 1);
+			long[] row = k2Raster.getWindow(i, i, 0, matrix.getWidth() - 1);
 			for (int j = 0; j < matrix.getWidth(); j++) {
 				assertEquals(matrix.get(i, j), row[j]);
 			}
@@ -202,8 +200,8 @@ public class KSquaredTest {
 		assertEquals(k2.k,k2_after.k);
 		assertEquals(k2.getSize(),k2_after.getSize());
 
-		int[] expected = k2.getWindow(0, k2.getSize()-1, 0, k2.getSize()-1);
-		int[] actual = k2_after.getWindow(0, k2.getSize()-1, 0, k2.getSize()-1);
+		long[] expected = k2.getWindow(0, k2.getSize()-1, 0, k2.getSize()-1);
+		long[] actual = k2_after.getWindow(0, k2.getSize()-1, 0, k2.getSize()-1);
 
 		for (int i = 0; i < expected.length; i++) {
 			assertEquals(expected[i], actual[i]);
